Fix thread-safety, DI design, and Error state in Insights cards
- Add @Volatile to isLoading/isLoadedSuccessfully flags in ViewModels
- Extract StatsSummaryUseCase singleton to replace summaryProvider var,
  removing temporal coupling and null-check error paths
- Change Error from class to data class, move onRetry out of state and
  into composable parameters to avoid unnecessary recompositions
