+++ WordPress/src/main/java/org/wordpress/android/ui/newstats/NewStatsActivity.kt
@@ -888,11 +888,6 @@ private fun InsightsTabContent(
         viewModel(),
     insightsViewModel: InsightsViewModel = viewModel()
 ) {
-    LaunchedEffect(Unit) {
-        val provider = insightsViewModel::getStatsSummary
-        allTimeStatsViewModel.summaryProvider = provider
-        mostPopularDayViewModel.summaryProvider = provider
-    }
     val context = LocalContext.current
     val yearInReviewUiState by yearInReviewViewModel
         .uiState.collectAsState()
@@ -1065,6 +1060,10 @@ private fun InsightsTabContent(
                                 insightsViewModel
                                     .removeCard(cardType)
                             },
+                            onRetry = {
+                                allTimeStatsViewModel
+                                    .onRetry()
+                            },
                             cardPosition = cardPosition,
                             onMoveUp = {
                                 insightsViewModel
@@ -1094,6 +1093,10 @@ private fun InsightsTabContent(
                                     .removeCard(
                                         cardType
                                     )
+                            },
+                            onRetry = {
+                                mostPopularDayViewModel
+                                    .onRetry()
                             },
                             cardPosition =
                                 cardPosition,
