+++ WordPress/src/main/java/org/wordpress/android/ui/newstats/mostpopularday/MostPopularDayCardUiState.kt
@@ -12,8 +12,7 @@ sealed class MostPopularDayCardUiState {
         val viewsPercentage: String
     ) : MostPopularDayCardUiState()
 
-    class Error(
-        val message: String,
-        val onRetry: () -> Unit
+    data class Error(
+        val message: String
     ) : MostPopularDayCardUiState()
 }
